requirements.txt
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.1
sentencepiece>=0.2.0
pyyaml>=6.0
tensorboard>=2.14.0
tqdm>=4.66.1
pandas>=2.0.0
sacrebleu>=2.3.1
scikit-learn>=1.3.0
